requirements.txt
PyQt5>=5.15.0
google-generativeai>=0.3.0
pandas>=1.3.0
openpyxl>=3.0.0
Pillow>=9.0.0
pyinstaller>=6.0.0
